requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
click==8.0.4
colorama==0.4.4
coverage==6.3.2
Django==4.0.4
django-autoslug==1.9.8
factory-boy==3.2.1
Faker==13.2.0
iniconfig==1.1.1
livereload==2.5.2
mypy-extensions==0.4.3
mysqlclient==2.1.0
packaging==21.3
parameterized==0.8.1
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.1.1
tzdata==2021.5
